fix: move remaining channel send-deps imports
