+++ src/cli/outbound-send-deps.ts
@@ -1,4 +1,4 @@
-import type { OutboundSendDeps } from "../infra/outbound/deliver.js";
+import type { OutboundSendDeps } from "../infra/outbound/send-deps.js";
 import {
   createOutboundSendDepsFromCliSource,
   type CliOutboundSendSource,
